comparmentalize rdf parsers
