--- components/niagads/ontology_parsers/__init__.py
@@ -1,3 +1,0 @@
-from niagads.ontology_parsers.owl import OWLParser
-
-__all__ = ["OWLParser"]
