+++ components/niagads/rdf_parsers/__init__.py
@@ -1,0 +1,3 @@
+from niagads.rdf_parsers.owl import OWLParser
+
+__all__ = ["OWLParser"]
